Add "icon-width" key to change the space between icons in Icon View.
The spacing between the icons and labels in a regular window
is quite large and might be a bit annoying when someone has search manually.

That space should be enough to fit an icon w/ label or two.

This adds predefined values aside from the defaults to manage the
grid where each file/folder or icons in general are placed. And, it also
adds a key or option to switch between the values at will.

*The original patch adds 3 options, this version condenses them to one.