icon-view: Add an option to change the space between the icons [WIP]
The spacing between the icons and labels in a regular window
is quite large and might be a bit annoying when someone
has search manually through a ton of files and folders .

That space should be enough to fit an icon, with its label, or two.

This adds 4 predefined values aside from the default to manage the
"grid" where each file/folder or icons in general are placed.
And, it also adds a schema key to switch between the values at will.
It shall be called as "icon-width" and should reside in org.nemo.icon-view.

Implements #802.